+++ .gitignore
@@ -7,3 +7,5 @@ __pycache__/
 
 .env
 os
+
+logs/*.json
