+++ backend/app.py
@@ -1,5 +1,5 @@
 from flask import Flask, request, jsonify
-from topstepx_trader.accounts import search_accounts
+from backend.topstepx_trader.accounts import search_accounts
 import os
 
 app = Flask(__name__)
