+++ backend/topstepx_trader/retrieve_bars.py
@@ -4,7 +4,7 @@ import requests
 import json
 from datetime import datetime, timedelta
 from dotenv import load_dotenv
-from topstepx_trader.contracts import search_contracts
+from backend.topstepx_trader.contracts import search_contracts
 
 load_dotenv()
