+++ config/jest.config.js
@@ -33,6 +33,7 @@ const config = {
 
   // Indicates which provider should be used to instrument code for coverage
   coverageProvider: "v8",
+  rootDir: "..",
   roots: ["<rootDir>/tests/js"],
 
   // A list of reporter names that Jest uses when writing coverage reports
