+++ frontend/renderer/onboarding.js
