+++ logs/contracts_log.json
