+++ package.json
@@ -2,10 +2,10 @@
   "name": "quantx-platform",
   "version": "0.1.0",
   "description": "QuantX Electron Frontend",
-  "main": "main.js",
+  "main": "electron/main.js",
   "scripts": {
     "start": "electron .",
-    "test": "jest",
+    "test": "jest --config config/jest.config.js",
     "dev": "vite",
     "build": "vite build"
   },
